Moved JSON utils to twc module, various cleanup
