+++ core/core.gradle
@@ -13,7 +13,6 @@ dependencies {
     api commonDependencies.'log4j'
     api commonDependencies.'log4j-api'
     api commonDependencies.'swagger-annotations'
-    api 'org.json:json:20190722'
 
     testImplementation commonDependencies.'spring-boot-starter-test'
 }
