+++ twc/src/main/java/org/openmbee/sdvc/twc/TeamworkCloudEndpoints.java
@@ -2,8 +2,8 @@ package org.openmbee.sdvc.twc;
 
 public enum TeamworkCloudEndpoints {
     LOGIN("login"),
-	GETROLESID("resources/%s/roles"),
-	GETPROJECTUSERS("workspaces/%s/resources/%s/roles/%s/users");
+    GETROLESID("resources/%s/roles"),
+    GETPROJECTUSERS("workspaces/%s/resources/%s/roles/%s/users");
 
     private String path;
