+++ twc/src/main/java/org/openmbee/sdvc/twc/utilities/TwcPermissionUtils.java
@@ -1,15 +1,12 @@
 package org.openmbee.sdvc.twc.utilities;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import org.openmbee.sdvc.core.utils.RestUtils;
-import org.openmbee.sdvc.core.utils.JsonUtils;
 import org.openmbee.sdvc.twc.TeamworkCloud;
 import org.openmbee.sdvc.twc.TeamworkCloudEndpoints;
-import org.openmbee.sdvc.twc.config.TwcConfig;
 import org.openmbee.sdvc.twc.constants.TwcConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
